Age-length input data to gadget

Age-length input data to gadget

Usage

aldist_example

Format

Data frame with gadget attributes containing Greenland halibut age-length distributions from the Institute of Marine Research EggaNord survey.

Source

Joint Russian-Norwegian Arctic Fisheries Working Group

See also