Plot of data passed into a gadget model

Plot all observational data from a gadget model

Usage

g3d_plot(model, regexp = NULL, scales = "fixed", ncol = NULL, base_size = 8)

Arguments

model

R or TMB model. A model object from g3_to_r or g3_to_tmb functions.

regexp

Either NULL (no filtering) or a character string giving a regular expression to filter model-data components. Useful alternatives: "adist" for abundance distribution data, "cdist" for catch distribution data, "surveyindices" for survey indices, and "catch" for catches.

scales

Character defining the facet_wrap scales argument to use.

ncol

Number of columns passed to facet_wrap

base_size

Base size parameter for ggplot. See ggtheme.

Value

A list of ggplot objects.

Details

Plots data as contained in a gadget model. Helpful in checking models for possible data issues and for documentation of models.