Plot gadget.fit object

A wrapper function to plot the results from a gadget.fit object.

Usage

# S3 method for class 'gadget.fit'
plot(x, param = "annual", ...)

Arguments

x

A gadget fit object. See g3_fit.

param

Character defining the parameter to plot. See Details.

...

Additional parameters passed to the separate plotting functions.

Value

Single or a list of ggplot objects depending on the arguments.

Details

A wrapper to plot the results from a gadget.fit object. The function produces a different plots defined by param argument and additional arguments passed to the respective functions.

Valid param options are:

annual

Annual summary of central model data. Uses plot_annual

biomass

Stock biomass. Uses plot_biomass

total

Stock biomass with total. Uses plot_biomass(..., total = TRUE)

abundance

Stock abundance. Uses plot_biomass(..., biomass = FALSE)

ssb

Spawning stock biomass. Uses plot_ssb

rec

Recruitment. Uses plot_rec

f

Fishing mortality. Uses plot_f

catch

Catches by stock. Uses plot_catch(..., type = "stock")

catch

Catches by fleet. Uses plot_catch(..., type = "fleet")

hr

Harvest rate. Uses plot_catch(..., type = "hr")

si

Survey indices. Uses plot_si

catchdist

Catch distribution comparison to data. Uses plot_catchdist

stockdist

Stock distribution comparison to data. Uses plot_stockdist

stockcomp

Model stock composition. Uses plot_stockcomp

suitablity

Suitability (fleet selection). Uses plot_suitability

growth

Average length by age. Uses plot_growth

agelength

Age-length (growth parameter) fit. Uses plot_agelength

agecomp

Age composition. Uses plot_agecomp

resid

Residuals. Uses plot_resid

likelihood

Likelihood scores. Uses plot_likelihood

weighted

Weighted likelihood scores. Uses plot_likelihood(..., type = "weighted")

pie

Proportion of summed weighted likelihood scores. Uses plot_likelihood(..., type = "pie")

weight

Parameter component weights

params

Parameter values relative to their boundaries